Branch Cleanup (#6)
* MapWidget Cleanup

Removed a bunch of old comments and backed out a couple test changes.  The page still intermittently fails to import the external google api, but it's a lot less frequent now.

* MapWidget Cleanup (#3) (#4)

Removed a bunch of old comments and backed out a couple test changes.  The page still intermittently fails to import the external google api, but it's a lot less frequent now.